# contracts/p5-reporting-decommission/openapi.yaml
openapi: 3.1.0
info:
  title: HRMS Phase 5 – Reporting / Integration, reference-data ownership, Forms + Oracle decommission
  version: 5.0.0-contract
  description: |
    Frozen API surface for Phase 5 (CUTOVER_PLAN.md §9, COMPONENT_MAPPING.md §9 rows
    `PKG_REPORTING` / `PKG_INTEGRATION`, MODERNIZATION_BLUEPRINT.md §8 Reporting row,
    TEST_STRATEGY.md §5 row 5). Written BEFORE any implementation exists; the backend
    node implements exactly this file, the frontend node codes against it.

    Target architecture (non-negotiable, MODERNIZATION_BLUEPRINT.md §9 decisions A/B):
    * Spring Boot (Java 21) on PostgreSQL + React. **No PL/SQL, PL/pgSQL, triggers or
      .pll in the target.** `PKG_REPORTING`, `PKG_INTEGRATION`, the six `VW_*` views and
      `HRMS_ADMIN.fmb` / `HRMS_REPORTS.fmb` are read-only characterization references
      (GOLDEN-ORACLE MODE = OFF: no Oracle runtime exists; expectations are recorded
      fixtures, `legacy_source=recorded`).
    * Three Java modules are introduced (ARCH-01/02, one owner per table):
      - `reporting-module` – **owns no table**. Every report is a read-only Spring Data
        JDBC projection (plain SQL, `WITH RECURSIVE` for the hierarchy) over tables owned
        by the P0–P4 modules, executed on the PostgreSQL read replica when one exists.
        The six `VW_*` views are **not** recreated on PostgreSQL; the report SQL is the
        `tests/reconciliation/pg/*.sql` transcription (Level 3 pins it).
...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/integration/gl-feed:
    post:
      tags: [integration]
      operationId: integrationGenerateGlFeed
      summary: "`PKG_INTEGRATION.generate_gl_journal(p_run_id)` → object storage"
      description: |
        Synchronous (a run is ≤ a few thousand `PAYROLL_DETAILS` rows). Writes
        `GL_JOURNAL_{runId}_{yyyyMMdd}.txt`, pipe-delimited:
        `H|HRMS_PAYROLL|{payDate:yyyy-MM-dd}|{runId}`, one
        `D|{costCenter}|{glAccount}|{debit}|{credit}|{description}|RUN-{runId}` per
        `cost_center × gl_account` (EARNING → `debit`, DEDUCTION/TAX → `credit`,
        magnitudes with two decimals, other side `0.00`), `T|{detailCount}`. `glAccount`
        comes from `PAY_ELEMENTS.GL_ACCOUNT`, `costCenter` from `DEPARTMENTS.COST_CENTER`
        (`UNASSIGNED` when null – legacy emitted an empty field; documented divergence).
        Only `APPROVED`/`PAID` runs are exportable (`-20701`). Re-running for the same
        run creates a **new** file version (legacy overwrote); `INTEGRATION_LOG` row
        `GL_JOURNAL` `SUCCESS|FAILED` with `created_by = jwt.sub`.
      x-preauthorize: hasAuthority('PAYROLL:APPROVE')
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GlFeedRequest' }
      responses:
        '201':
          description: Feed written
          headers:
            Location: { schema: { type: string }, description: "`/api/integration/files/{fileId}`" }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationFile' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '422':
          $ref: '#/components/responses/RunNotExportable'
  /api/integration/benefits-feed:
    post:
      tags: [integration]
      operationId: integrationExportBenefitsFeed
      summary: "`PKG_INTEGRATION.export_benefits_feed(p_effective_date)` → object storage"
      description: |
        Fixed-width ADP layout: one `E` record per active employee (`empNumber` 10,
        `lastName` 30, `firstName` 30, `ssnMasked` 11, `dateOfBirth` yyyyMMdd, `gender`
        1, `hireDate` yyyyMMdd, `maritalStatus` 1, `baseSalary` 12.2, `deptCode` 20,
        `locationCode` 10) followed by one `D` record per `EMPLOYEE_DEPENDENTS` row with
        `IS_BENEFICIARY='Y'` (`relationship` 20, `lastName` 30, `firstName` 30,
        `dateOfBirth` yyyyMMdd, `ssnMasked` 11). **SSN is masked `***-**-dddd`** (SEC-05;
        the legacy feed carried the full SSN – intentionally not reproduced). File name
        `BENEFITS_{yyyyMMdd}.txt`. `INTEGRATION_LOG` row `BENEFITS_FEED`.
      x-preauthorize: hasAuthority('ADMIN:EDIT') and hasAuthority('EMPLOYEE:VIEW')
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BenefitsFeedRequest' }
      responses:
        '201':
          description: Feed written
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationFile' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/integration/time-attendance/import:
    post:
      tags: [integration]
      operationId: integrationImportTimeAttendance
      summary: "`PKG_INTEGRATION.import_time_attendance` – BUG-08: validate + stage only"
      description: |
        `multipart/form-data` upload (`file`, ≤ 5 MiB, `text/csv`) replacing the
        `TIME_IMPORT_IN` directory read. Grammar frozen from the legacy comment:
        `emp_number,date,hours_regular,hours_overtime`, optional header row, `date` =
        `YYYY-MM-DD`, hours `0.00`–`24.00` with ≤ 2 decimals, `hours_regular +
        hours_overtime <= 24`. Per-line verdicts: `ACCEPTED`, or `REJECTED` with
        `ApiError` (`-20001` unknown/inactive `emp_number`; `VALIDATION_FAILED` bad
        date/hours; `-20703` duplicate `(emp_number, date)` within the file). The whole
        file is rejected with `415` on a wrong content type and `422` (`-20704`) when
        **no** line is accepted.

        **BUG-08 – `x-unspecified`:** the legacy procedure never wrote anything (its loop
        body is `-- TODO`). The target **does not invent** the persistence: accepted rows
        are staged (`integration_files.status = STAGED`, rows kept as the file content)
        and **no `PAYROLL_*`, `EMPLOYEE_PAY_ELEMENTS` or `LEAVE_*` row is touched**. The
        mapping of `hours_regular`/`hours_overtime` to pay elements and the target
        table are marked `x-unspecified: BUG-08` on `TimeAttendanceImportResult` and
        require a product decision before any later phase mounts a `POST
        /api/integration/time-attendance/{fileId}/apply`. Overtime eligibility
        (`JOB_GRADES.OVERTIME_ELIGIBLE`) is **reported** per line (`overtimeEligible`)
        but not enforced, for the same reason.
      x-preauthorize: hasAuthority('PAYROLL:EDIT')
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "`text/csv`, UTF-8, ≤ 5 MiB, ≤ 50 000 lines"
                hasHeader:
                  type: boolean
                  default: true
      responses:
        '201':
          description: Staged (at least one line accepted)
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeAttendanceImportResult' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: File larger than 5 MiB
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example: { code: PAYLOAD_TOO_LARGE, message: File exceeds 5 MiB, field: file, traceId: 4f3c… }
        '415':
          description: Not `text/csv`
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
              example: { code: UNSUPPORTED_MEDIA_TYPE, message: Expected text/csv, field: file, traceId: 4f3c… }
        '422':
          $ref: '#/components/responses/ImportRejected'
  /api/integration/files:
    get:
      tags: [integration]
      operationId: integrationListFiles
      summary: Paged feed / import file metadata (`INTEGRATION_LOG` joined)
      x-preauthorize: hasAuthority('ADMIN:VIEW')
      parameters:
        - in: query
          name: feed
          required: false
          schema: { $ref: '#/components/schemas/FeedType' }
        - in: query
          name: status
          required: false
          schema: { $ref: '#/components/schemas/IntegrationFileStatus' }
        - in: query
          name: from
          required: false
          schema: { type: string, format: date }
        - in: query
          name: to
          required: false
          schema: { type: string, format: date }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: Files
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationFilePage' }
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/integration/files/{fileId}:
    get:
      tags: [integration]
      operationId: integrationGetFile
      summary: Metadata of one file
      x-preauthorize: hasAuthority('ADMIN:VIEW')
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: File metadata
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IntegrationFile' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/FileNotFound'
  /api/integration/files/{fileId}/content:
    get:
      tags: [integration]
      operationId: integrationDownloadFile
      summary: Stream the artefact from object storage (`UTL_FILE` directory replacement)
      description: |
        `Content-Disposition: attachment; filename="{fileName}"`, `Content-Type` =
        `text/plain; charset=utf-8` for GL/benefits, `text/csv` for imports, `ETag` =
        `sha256`. GL files require `PAYROLL:VIEW`, benefits files `EMPLOYEE:VIEW`
        (masked SSN notwithstanding), imports `PAYROLL:VIEW`; `ADMIN:VIEW` alone gets
        the metadata but a `403` here.
      x-preauthorize: "hasAuthority('ADMIN:VIEW') and @integrationAccess.canDownload(#fileId, authentication)"
      parameters:
        - $ref: '#/components/parameters/FileIdPath'
      responses:
        '200':
          description: File bytes
          headers:
            Content-Disposition: { schema: { type: string } }
            ETag: { schema: { type: string } }
          content:
            text/plain:
              schema: { type: string }
            text/csv:
              schema: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/FileNotFound'
  /api/integration/status:
    get:
      tags: [integration]
      operationId: integrationStatus
      summary: "`PKG_INTEGRATION.get_integration_status` for every feed"
      description: Latest `INTEGRATION_LOG` row per feed (`status`, `lastRunAt`, `lastFileId`, `message`); `NEVER_RUN` when none.
      x-preauthorize: hasAuthority('ADMIN:VIEW')
      responses:
        '200':
          description: Status per feed
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/IntegrationStatus' }
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: P0 access token; claims `sub` (user-account id, string), `empId`, `authorities[]`.
  parameters:
...
    Page:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      in: query
      name: size
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    FileIdPath:
      in: path
      name: fileId
      required: true
      schema: { type: string, format: uuid }
  responses:
...
    ValidationFailed:
      description: Bean Validation / unsupported query value (`validation-schema.json` module `p5-reporting-decommission`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: VALIDATION_FAILED, message: asOf must be a date (YYYY-MM-DD), field: asOf, traceId: 4f3c… }
    Unauthorized:
      description: Missing / invalid / revoked token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: TOKEN_INVALID, message: Access token is missing or invalid, traceId: 4f3c… }
    Forbidden:
      description: "`@PreAuthorize` denied (`REPORTS:VIEW`, `PAYROLL:VIEW`, `ADMIN:VIEW`, `ADMIN:EDIT`, `LEAVE:ADMIN`, `PAYROLL:APPROVE`, `PAYROLL:EDIT`, `EMPLOYEE:VIEW`)"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: FORBIDDEN, message: Access denied, traceId: 4f3c… }
...
    RunNotFound:
      description: "`RUN_NOT_FOUND`: `runId` matches no `PAYROLL_RUNS` row (P4 framework code)"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: RUN_NOT_FOUND, message: 'Payroll run not found: 1001', field: runId, traceId: 4f3c… }
...
    RunNotExportable:
      description: "`-20701`: `PAYROLL_RUNS.STATUS` is not `APPROVED` or `PAID`"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: '-20701', message: 'Cannot export GL feed for run in status: CALCULATED', field: runId, traceId: 4f3c… }
    ImportRejected:
      description: "`-20704`: no line of the time-attendance file was accepted; `details` carries the first 100 line errors"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example:
            code: '-20704'
            message: 'No valid rows in time-attendance file (12 lines rejected)'
            field: file
            traceId: 4f3c…
            details: [{ field: 'line 2', message: '-20001 Employee not found or not active: E9999' }]
    FileNotFound:
      description: "`FILE_NOT_FOUND`: unknown `fileId` (metadata) or the object is gone from storage"
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
          example: { code: FILE_NOT_FOUND, message: 'Integration file not found', traceId: 4f3c… }
  schemas:
    ApiError:
      type: object
      required: [code, message, traceId]
      additionalProperties: false
      properties:
        code:
          type: string
          description: Legacy `-20xxx` as a string, or a framework code (`error-codes.md`).
        message: { type: string }
        field:
          type: [string, 'null']
          description: Offending request field (body property, query parameter, multipart part or `line N`) when attributable.
        traceId:
          type: string
          description: Correlates with `error_log.trace_id` / `X-Trace-Id`.
        details:
          type: array
          description: All violations for `VALIDATION_FAILED` / `-20704`.
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }
...
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        totalElements: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    FeedType:
      type: string
      enum: [GL_JOURNAL, BENEFITS_FEED, TIME_ATTENDANCE]
    IntegrationFileStatus:
      type: string
      enum: [SUCCESS, FAILED, STAGED]
      description: "`STAGED` only for time-attendance imports (BUG-08)."
    GlFeedRequest:
      type: object
      required: [runId]
      additionalProperties: false
      properties:
        runId: { type: integer, format: int64, minimum: 1, description: "`PAYROLL_RUNS.RUN_ID`, status APPROVED/PAID." }
    BenefitsFeedRequest:
      type: object
      additionalProperties: false
      properties:
        effectiveDate: { type: string, format: date, description: "Defaults to today (`p_effective_date DEFAULT SYSDATE`); selects active employees hired on/before it." }
    IntegrationFile:
      type: object
      required: [fileId, feed, fileName, status, sizeBytes, sha256, recordCount, createdBy, createdAt, contentUrl]
      properties:
        fileId: { type: string, format: uuid }
        feed: { $ref: '#/components/schemas/FeedType' }
        fileName: { type: string, example: GL_JOURNAL_1001_20240705.txt }
        status: { $ref: '#/components/schemas/IntegrationFileStatus' }
        sizeBytes: { type: integer, minimum: 0 }
        sha256: { type: string, pattern: '^[0-9a-f]{64}$' }
        recordCount: { type: integer, minimum: 0, description: "`D` records / `E`+`D` records / accepted lines." }
        sourceRef: { type: [string, 'null'], description: "`runId` for GL, `effectiveDate` for benefits, original upload name for imports." }
        storageKey: { type: string, description: "`{feed}/{yyyy}/{MM}/{fileId}-{fileName}` in bucket `hrms-integration`." }
        contentUrl: { type: string, description: "`/api/integration/files/{fileId}/content`" }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
        message: { type: [string, 'null'] }
    IntegrationFilePage:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/IntegrationFile' }
        page: { $ref: '#/components/schemas/PageMeta' }
    TimeAttendanceLine:
      type: object
      required: [line, empNumber, workDate, hoursRegular, hoursOvertime, verdict]
      properties:
        line: { type: integer, minimum: 1 }
        empNumber: { type: string }
        empId: { type: [integer, 'null'], format: int64 }
        workDate: { type: string, format: date }
        hoursRegular: { type: string, pattern: '^[0-9]{1,2}\.[0-9]{2}$' }
        hoursOvertime: { type: string, pattern: '^[0-9]{1,2}\.[0-9]{2}$' }
        overtimeEligible: { type: [boolean, 'null'], description: "`JOB_GRADES.OVERTIME_ELIGIBLE` of the employee's grade – reported, not enforced (BUG-08)." }
        verdict: { type: string, enum: [ACCEPTED, REJECTED] }
        error:
          oneOf:
            - $ref: '#/components/schemas/ApiError'
            - type: 'null'
    TimeAttendanceImportResult:
      type: object
      required: [file, accepted, rejected, lines, applied, targetTable, payElementMapping]
      properties:
        file: { $ref: '#/components/schemas/IntegrationFile' }
        accepted: { type: integer, minimum: 0 }
        rejected: { type: integer, minimum: 0 }
        lines:
          type: array
          items: { $ref: '#/components/schemas/TimeAttendanceLine' }
        applied:
          type: boolean
          const: false
          description: Always `false` in Phase 5 – rows are staged only.
          x-unspecified: BUG-08
        targetTable:
          type: 'null'
          description: The table the hours would be written to. **Unspecified by the legacy source** (no time/attendance table exists).
          x-unspecified: BUG-08
        payElementMapping:
          type: 'null'
          description: Which `PAY_ELEMENTS` rows `hours_regular` / `hours_overtime` map to. **Unspecified by the legacy source.**
          x-unspecified: BUG-08
    IntegrationStatus:
      type: object
      required: [feed, status]
      properties:
        feed: { $ref: '#/components/schemas/FeedType' }
        status: { type: string, enum: [NEVER_RUN, SUCCESS, FAILED, STAGED] }
        lastRunAt: { type: [string, 'null'], format: date-time }
        lastFileId: { type: [string, 'null'], format: uuid }
        lastRunBy: { type: [string, 'null'] }
        message: { type: [string, 'null'] }